+++ src/SeablastConfiguration.php
@@ -5,8 +5,7 @@ declare(strict_types=1);
 namespace Seablast\Seablast;
 
 use Seablast\Seablast\Exceptions\DbmsException;
-use Seablast\Seablast\SeablastConfigurationException;
-use Seablast\Seablast\SeablastMysqli;
+use Seablast\Seablast\Exceptions\SeablastConfigurationException;
 use Tracy\Debugger;
 use Webmozart\Assert\Assert;
 
@@ -17,12 +16,12 @@ class SeablastConfiguration
 {
     use \Nette\SmartObject;
 
-    /** @var ?SeablastMysqli */
-    private $connection = null;
     /** @var ?string */
     private $connectionTablePrefix = null;
     /** @var SeablastFlag */
     public $flag;
+    /** @var ?SeablastMysqli */
+    private $mysqli = null;
     /** @var array<array<string[]>> */
     private $optionsArrayArrayString = [];
     /** @var array<int[]> */
@@ -33,6 +32,10 @@ class SeablastConfiguration
     private $optionsInt = [];
     /** @var string[] */
     private $optionsString = [];
+    /** @var ?SeablastPdo */
+    private $pdo = null;
+    /** @var string|null for db logging */
+    private $user = null;
 
     public function __construct()
     {
@@ -42,26 +45,26 @@ class SeablastConfiguration
     /**
      * Access to database with lazy initialization.
      *
+     * todo alias for mysqli and this deprecated
+     *
      * @return SeablastMysqli
      */
     public function dbms(): SeablastMysqli
     {
         //Lazy initialisation
         if (!$this->dbmsStatus()) {
-            Debugger::barDump('Creating database connection');
-            $this->dbmsCreate();
+            Debugger::barDump('Creating MySQLi database connection');
+            $this->mysqliCreate();
         }
-        Assert::object($this->connection);
-        Assert::isAOf($this->connection, '\Seablast\Seablast\SeablastMysqli');
-        return $this->connection;
+        Assert::object($this->mysqli);
+        Assert::isAOf($this->mysqli, '\Seablast\Seablast\SeablastMysqli');
+        return $this->mysqli;
     }
 
     /**
-     * Creates a database connection and sets up charset.
-     *
-     * @return void
+     * Read phinx configuration and provide pertinent parameters in a type strict manner
      */
-    private function dbmsCreate(): void
+    private function dbmsExtractProperties(): DatabaseProperties
     {
         $phinx = self::dbmsReadPhinx();
         Assert::isArray($phinx['environments']);
@@ -81,26 +84,19 @@ class SeablastConfiguration
         } else {
             $port = null;
         }
-        $this->connection = new SeablastMysqli(
+        return new DatabaseProperties(
+            $phinx['environments'][$environment]['adapter'],
             $phinx['environments'][$environment]['host'], // todo fix localhost
             $phinx['environments'][$environment]['user'],
             $phinx['environments'][$environment]['pass'],
             $phinx['environments'][$environment]['name'],
-            $port
+            $port,
+            $phinx['environments'][$environment]['table_prefix'] ?? ''
         );
-        // todo does this really differentiate between successful connection, failed connection and no connection?
-        Assert::isAOf($this->connection, '\Seablast\Seablast\SeablastMysqli');
-        Assert::true(
-            $this->connection->set_charset($this->getString(SeablastConstant::SB_CHARSET_DATABASE)),
-            'Unexpected character set: ' . $this->getString(SeablastConstant::SB_CHARSET_DATABASE)
-        );
-        $this->connectionTablePrefix = $phinx['environments'][$environment]['table_prefix'] ?? '';
     }
 
     /**
      * Return the table prefix from phinx config.
-     *
-     * TODO experimental - keep only if working well
      *
      * @return string
      */
@@ -131,11 +127,13 @@ class SeablastConfiguration
      *
      * So that the SQL Bar Panel is not requested in vain.
      *
+     todo alias for mysqliStatus and deprecated
+
      * @return bool
      */
     public function dbmsStatus(): bool
     {
-        return $this->connection instanceof \mysqli;
+        return $this->mysqli instanceof \mysqli;
     }
 
     /**
@@ -232,6 +230,89 @@ class SeablastConfiguration
     }
 
     /**
+     * Creates a database connection and sets up charset.
+     *
+     * @return void
+     */
+    private function mysqliCreate(): void
+    {
+        $phinx = $this->dbmsExtractProperties();
+        $this->mysqli = new SeablastMysqli(
+            $phinx->host, // todo fix localhost
+            $phinx->user,
+            $phinx->pass,
+            $phinx->name,
+            $phinx->port
+        );
+        // todo does this really differentiate between successful connection, failed connection and no connection?
+        Assert::isAOf($this->mysqli, '\Seablast\Seablast\SeablastMysqli');
+        Assert::true(
+            $this->mysqli->set_charset($this->getString(SeablastConstant::SB_CHARSET_DATABASE)),
+            'Unexpected character set: ' . $this->getString(SeablastConstant::SB_CHARSET_DATABASE)
+        );
+        $this->connectionTablePrefix = $phinx->tablePrefix;
+        // if user already set then setUser
+        if (!is_null($this->user)) {
+            $this->mysqli->setUser($this->user);
+        }
+    }
+
+    /**
+     * Access to database using PDO adapter with lazy initialization.
+     *
+     * @return SeablastPdo
+     */
+    public function pdo(): SeablastPdo
+    {
+        //Lazy initialisation
+        if (!$this->pdoStatus()) {
+            Debugger::barDump('Creating PDO database connection');
+            $this->pdoCreate();
+        }
+        Assert::object($this->pdo);
+        Assert::isAOf($this->pdo, '\Seablast\Seablast\SeablastPdo');
+        return $this->pdo;
+    }
+
+    /**
+     * Creates a database connection and sets up charset.
+     *
+     * @return void
+     */
+    private function pdoCreate(): void
+    {
+        $phinx = $this->dbmsExtractProperties();
+        //PDO("mysql:host=localhost;dbname=DB;charset=UTF8")
+        $this->pdo = new SeablastPdo(
+            //   $phinx->host, // todo fix localhost
+            "{$phinx->adapter}:host={$phinx->host}"
+            . (is_null($phinx->port) ? '' : ";port={$phinx->port}")
+            . ";dbname={$phinx->name};charset={$this->getString(SeablastConstant::SB_CHARSET_DATABASE)}",
+            $phinx->user,
+            $phinx->pass
+        );
+        // todo does this really differentiate between successful connection, failed connection and no connection?
+        Assert::isAOf($this->pdo, '\Seablast\Seablast\SeablastPdo');
+        $this->connectionTablePrefix = $phinx->tablePrefix;
+        // if user already set then setUser
+        if (!is_null($this->user)) {
+            $this->pdo->setUser($this->user);
+        }
+    }
+
+    /**
+     * Returns true if connected, false otherwise.
+     *
+     * So that the SQL Bar Panel is not requested in vain.
+     *
+     * @return bool
+     */
+    public function pdoStatus(): bool
+    {
+        return $this->pdo instanceof \PDO;
+    }
+
+    /**
      * @param string $property
      * @param string $key
      * @param string[] $value
@@ -297,5 +378,40 @@ class SeablastConfiguration
     {
         $this->optionsString[$property] = $value;
         return $this;
+    }
+
+    /**
+     * DI setter for databases.
+     *
+     * @param int|string $user
+     *
+     * @return void
+     */
+    public function setUser($user): void
+    {
+        $this->user = (string) $user;
+        // if mysqli or pdo then setUser
+        if ($this->dbmsStatus()) {
+            $this->dbms()->setUser($this->user);
+        }
+        if ($this->pdoStatus()) {
+            $this->pdo()->setUser($this->user);
+        }
+    }
+
+    /**
+     * Show Tracy BarPanel with SQL statements.
+     *
+     * @return void
+     */
+    public function showSqlBarPanel(): void
+    {
+        // if mysqli or pdo then showSqlBarPanel
+        if ($this->dbmsStatus()) {
+            $this->dbms()->showSqlBarPanel();
+        }
+        if ($this->pdoStatus()) {
+            $this->pdo()->showSqlBarPanel();
+        }
     }
 }