+++ src/SeablastModel.php
@@ -39,7 +39,13 @@ class SeablastModel
             Assert::true(class_exists($className), "Class {$className} does not exist.");
             $model = new $className($this->controller->getConfiguration(), $superglobals);
             Assert::methodExists($model, 'knowledge', "{$className} model MUST have method knowledge()");
-            $this->viewParameters = $model->knowledge();
+            try {
+                $this->viewParameters = $model->knowledge();
+            } catch (Exceptions\DbmsException $e) {
+                // make sure that the database Tracy BarPanel is displayed when DbmsException is thrown
+                $this->controller->getConfiguration()->showSqlBarPanel();
+                throw new Exceptions\DbmsException($e->getMessage(), $e->getCode(), $e);
+            }
             Debugger::log('knowledge of ' . $className . ': ' . print_r($this->viewParameters, true), ILogger::DEBUG);
             Assert::isAOf($this->viewParameters, 'stdClass', "The knowledge of {$className} MUST be of stdClass type.");
         } else {